+++ CHANGES.md
@@ -1,5 +1,9 @@
 Markdown Writer FX Change Log
 =============================
 
+## 0.2
+- pegdown Markdown parser updated to version 1.6.
+- Added five new pegdown 1.6 extension flags to Markdown Options tab.
+
 ## 0.1
 - Initial release